Add hashcode and equals methods to LatLong (#179)
* Add hashcode and equals methods to LatLong

Discovered that LatLong doesn't actually implement hashcode() or
equals(), so we can't actually put it into a Set<> or call distinct() on
it in a stream.

Used IntelliJ IDEA's automatic generation of these things. Happy to
change the implementation if desired.

* Performance fix

Objects.hashCode() under the hood creates a bunch of objects we don't
need, so directly create and aggregate the hashcodes.